+++ sdk/schemaregistry/schema-registry-json/samples-dev/schemaRegistryJsonWithValidation.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 /**
- * @summary Demonstrates the use of JsonSerializer to create messages with json-serialized payload using schema from Schema Registry with validation using a third party library.
+ * @summary Demonstrates the use of JsonSchemaSerializer to create messages with json-serialized payload using schema from Schema Registry with validation using a third party library.
  */
 
 import { DefaultAzureCredential } from "@azure/identity";
@@ -11,7 +11,7 @@ import {
   SchemaDescription,
   KnownSchemaFormats,
 } from "@azure/schema-registry";
-import { DeserializeOptions, JsonSerializer } from "@azure/schema-registry-json";
+import { DeserializeOptions, JsonSchemaSerializer } from "@azure/schema-registry-json";
 
 import Ajv, { ValidateFunction } from "ajv";
 // Load the .env file if it exists
@@ -69,7 +69,7 @@ export async function main() {
   await client.registerSchema(schemaDescription);
 
   // Create a new serializer backed by the client
-  const serializer = new JsonSerializer(client, { groupName });
+  const serializer = new JsonSchemaSerializer(client, { groupName });
 
   // serialize an object that matches the schema and put it in a message
   const value: User = { firstName: "Jane", lastName: "Doe" };