+++ sdk/schemaregistry/schema-registry-json/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Other Changes
 
+- Update `JsonSerializer` class name to `JsonSchemaSerializer`
+
 ## 1.0.0-beta.1 (2023-08-08)
 
 - Initial preview release